refactor(test): move ISEMAIL tests to tests/ subdirectory pattern
Removes inline test modules from is_checks/mod.rs and adds
tests/isemail.rs following the project test layout (tests/mod.rs +
submodule files). Tests the private is_valid_email helper and the
ISEMAIL function via crate::evaluate().
