+++ crates/core/src/eval/functions/logical/is_checks/mod.rs
@@ -203,86 +203,3 @@ fn is_date_string(s: &str) -> bool {
     (1..=12).contains(&month) && (1..=31).contains(&day)
 }
 
-#[cfg(test)]
-mod isemail_tests {
-    use super::is_valid_email;
-
-    // ── Success cases ─────────────────────────────────────────────────────────
-
-    #[test]
-    fn valid_simple_email() {
-        assert!(is_valid_email("user@example.com"));
-    }
-
-    #[test]
-    fn valid_with_subdomain() {
-        assert!(is_valid_email("user@mail.example.com"));
-    }
-
-    #[test]
-    fn valid_with_plus() {
-        assert!(is_valid_email("user+tag@example.com"));
-    }
-
-    // ── Edge cases ────────────────────────────────────────────────────────────
-
-    #[test]
-    fn missing_at_sign() {
-        assert!(!is_valid_email("notanemail"));
-    }
-
-    #[test]
-    fn missing_domain() {
-        assert!(!is_valid_email("user@"));
-    }
-
-    #[test]
-    fn missing_local_part() {
-        assert!(!is_valid_email("@example.com"));
-    }
-
-    #[test]
-    fn no_dot_in_domain() {
-        // "user@localhost" has no dot — domain must have at least one dot
-        assert!(!is_valid_email("user@localhost"));
-    }
-}
-
-#[cfg(test)]
-mod isemail_fn_tests {
-    use super::isemail_fn;
-    use crate::eval::functions::EvalCtx;
-    use crate::eval::{Context, Registry};
-    use crate::parser::ast::{Expr, Span};
-    use crate::types::{ErrorKind, Value};
-
-    fn span() -> Span { Span::new(0, 1) }
-
-    fn run_isemail(expr: Expr) -> Value {
-        let registry = Registry::new();
-        let ctx = Context::empty();
-        let mut eval_ctx = EvalCtx::new(ctx, &registry);
-        isemail_fn(&[expr], &mut eval_ctx)
-    }
-
-    #[test]
-    fn number_arg_returns_false() {
-        let result = run_isemail(Expr::Number(42.0, span()));
-        assert_eq!(result, Value::Bool(false));
-    }
-
-    #[test]
-    fn bool_arg_returns_false() {
-        let result = run_isemail(Expr::Bool(true, span()));
-        assert_eq!(result, Value::Bool(false));
-    }
-
-    #[test]
-    fn zero_args_returns_error_na() {
-        let registry = Registry::new();
-        let ctx = Context::empty();
-        let mut eval_ctx = EvalCtx::new(ctx, &registry);
-        let result = isemail_fn(&[], &mut eval_ctx);
-        assert_eq!(result, Value::Error(ErrorKind::NA));
-    }
-}
