+++ crates/core/src/eval/functions/logical/is_checks/tests/mod.rs
@@ -1,3 +1,4 @@
 mod success;
 mod failure;
 mod edge;
+mod isemail;
